Add files via upload
Unfortunately, I still get errors when trying to create the environment file, both via conda or the command prompt (probably due to the blank in the file path which I unfortunately can not change as it is in my user name). I will give it a try at the computers at uni and see if it works there. Until then I can not provide the environment file. However, the main library is installed using the pip command in the file, any other missing libraries can easily be added to the pip command line.